Struct ml_rust_cuda::math::linear::Vector

pub struct Vector { /* fields omitted */ }

A representation of a 32-bit float mathematical vector.

A vector is simply a wrapper around a Matrix with additional and/or vector-specific operations. Don't confuse this with Rust's Vec data type.

use ml_rust_cuda::math::linear::Vector;

let elements = vec![1_f32, 3_f32, 4_f32];
let vector = Vector::new(elements);

// "Vector { matrix: Matrix { elements: [ 1.0, 3.0, 4.0 ], dims: (3, 1) } }
println!("{:?}", vector);

Implementations

impl Vector

pub fn new(elements: Vec<f32>) -> Self

Returns a mathematical vector representation from a supplied f32 vector.

Arguments

elements - A vector of f32s for the mathematical vector.

Examples

use ml_rust_cuda::math::linear::Vector;

let vec = vec![9_f32, 5_f32, 8_f32, 3_f32, 4_f32, 0_f32];
let vector = Vector::new(vec);

println!("{}", vector);

pub fn from_matrix(matrix: Matrix) -> Self

Returns a mathematical vector around a given Matrix.

Arguments

matrix - The matrix to contain in the resulting vector.

Examples

use ml_rust_cuda::math::linear::{ Matrix, Vector };

let matrix = Matrix::new(vec![vec![1_f32, 2_f32, 5_f32, 3_f32]]);
let vector = Vector::from_matrix(matrix);

println!("{}", vector);

pub fn zero(n_elements: usize) -> Self

Returns a mathematical vector of specified dimension filled with zeros.

Arguments

n_elements - The dimension of the resulting vector.

Examples

use ml_rust_cuda::math::linear::Vector;

let vector = Vector::zero(6);

println!("{}", vector);

pub fn get(&self, i: usize) -> Option<f32>

Returns the element of the vector at a given position, or None if the index is out of bounds of the vector.

Arguments

i - The target element's index.

Examples

use ml_rust_cuda::math::linear::Vector;

let vec = vec![9_f32, 5_f32, 8_f32, 3_f32, 4_f32, 0_f32];
let vector = Vector::new(vec);

assert_eq!(vector.get(2), Some(8_f32));
assert!(vector.get(8).is_none());

pub fn set(&mut self, i: usize, val: f32)

Replaces the element of the vector at a given position with a given value. Does nothing if the position is out of bounds of the vector.

Arguments

i - The target element's position.

val - The value with which to replace the target element.

Examples

use ml_rust_cuda::math::linear::Vector;

let vec = vec![9_f32, 5_f32, 8_f32, 3_f32, 4_f32, 0_f32];
let mut vector = Vector::new(vec);

vector.set(2, 5_f32);

assert_eq!(vector.get(2), Some(5_f32));

pub fn matrix(&self) -> &Matrix

Returns a reference to the interal matrix ofthe vector.

pub fn dim(&self) -> usize

Returns the dimension of the vector.

pub fn dot(&self, other: &Self) -> f32

Returns the dot product of the vector with another vector.

Uses a CUDA kernel under the hood.

Arguments

other - A reference to another vector operand.

Examples

use ml_rust_cuda::math::linear::Vector;

let v1 = Vector::new(vec![9_f32, 5_f32, 8_f32, 3_f32, 4_f32, 0_f32]);
let v2 = Vector::new(vec![8_f32, 1_f32, 2_f32, 4_f32, 0_f32, 1_f32]);

let dot = v1.dot(&v2);
assert_eq!(dot, 105_f32);

Panics

Panics if the dimensions of both vectors are not equal.

pub fn p_norm(&self, p: f32) -> f32

Returns the p-norm of the vector.

The p-norm of a vector v of dimension n is defined as

||v||p = (|v_1|^p + ... |v_n|^p)^(1/p)

for any real number p >= 0.

The infinity norm (i.e. the limit of ||v||p as p tends to infinity) is the same as max(v_i) for each element v_i of v.

Uses a CUDA kernel under the hood.

Arguments

p - A positive 32-bit float.

Examples

use ml_rust_cuda::math::linear::Vector;

let v = Vector::new(vec![9_f32, 5_f32, 8_f32, 3_f32, 4_f32, 0_f32]);
let p = 2_f32;

let p_norm = v.p_norm(p);
let inf_norm = v.p_norm(f32::INFINITY);
assert_eq!(p_norm, 195_f32.sqrt());
assert_eq!(inf_norm, 9_f32);

Panics

Panics if p is less than or equal to 0.